requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
facenet-pytorch==2.5.2
idna==3.2
numpy==1.21.2
opencv-python==4.5.3.56
piexif==1.1.3
Pillow==8.3.2
requests==2.26.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
typing-extensions==3.10.0.2
urllib3==1.26.6